refactor(source-mssql): extract engine-independent e2e harness into db-harness-lib
